chore: move selectors to their own file
